Move `XArrayNetCDFCodec` into `utils/codec.py`
- drop `utils/xarray_codec.py`; all three codecs now live in one module
- fold the module docstring into the class docstring, shortened
- single codec import in `dj_pipeline/__init__.py`
