+++ aeon/dj_pipeline/__init__.py
@@ -39,12 +39,10 @@ pymysql.converters.escape_dict = _escape_dict_as_json
 pymysql.converters.encoders[dict] = _escape_dict_as_json
 pymysql.converters.conversions[dict] = _escape_dict_as_json
 
-# Register Aeon codecs BEFORE any schema activation
+# Register Aeon + xarray codecs BEFORE any schema activation
 from aeon.dj_pipeline.utils.codec import (  # pyright: ignore[reportUnusedImport]
     AeonStreamCodec,
     OnixStreamCodec,
-)
-from aeon.dj_pipeline.utils.xarray_codec import (  # pyright: ignore[reportUnusedImport]
     XArrayNetCDFCodec,
 )
